--- development/tools/jobs/incubator/documentation-component/governance_test.go
@@ -1,59 +1,0 @@
-package documentation_component_test
-
-import (
-	"testing"
-
-	"fmt"
-
-	"github.com/kyma-project/test-infra/development/tools/jobs/tester"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
-
-func TestGovernanceJobPresubmit(t *testing.T) {
-	// WHEN
-	jobConfig, err := tester.ReadJobConfig("./../../../../../prow/jobs/incubator/documentation-component/documentation-component-governance.yaml")
-	// THEN
-	require.NoError(t, err)
-
-	expName := "pre-master-documentation-component-governance"
-	actualPresubmit := tester.FindPresubmitJobByName(jobConfig.Presubmits["kyma-incubator/documentation-component"], expName, "master")
-	require.NotNil(t, actualPresubmit)
-	assert.Equal(t, expName, actualPresubmit.Name)
-	assert.Equal(t, []string{"^master$"}, actualPresubmit.Branches)
-	assert.Equal(t, 10, actualPresubmit.MaxConcurrency)
-	assert.False(t, actualPresubmit.SkipReport)
-	assert.True(t, actualPresubmit.Decorate)
-	assert.Equal(t, "github.com/kyma-incubator/documentation-component", actualPresubmit.PathAlias)
-	tester.AssertThatHasExtraRefTestInfra(t, actualPresubmit.JobBase.UtilityConfig, "master")
-	tester.AssertThatHasPresets(t, actualPresubmit.JobBase, tester.PresetBuildPr, tester.PresetDindEnabled)
-	assert.Equal(t, "milv.config.yaml|.md$", actualPresubmit.RunIfChanged)
-	tester.AssertThatJobRunIfChanged(t, *actualPresubmit, "milv.config.yaml")
-	tester.AssertThatJobRunIfChanged(t, *actualPresubmit, "some_markdown.md")
-	assert.Equal(t, tester.ImageBootstrapLatest, actualPresubmit.Spec.Containers[0].Image)
-	assert.Equal(t, []string{tester.GovernanceScriptDir}, actualPresubmit.Spec.Containers[0].Command)
-	assert.Equal(t, []string{"--repository", "documentation-component", "--repository-org", "kyma-incubator"}, actualPresubmit.Spec.Containers[0].Args)
-}
-
-func TestGovernanceJobPeriodic(t *testing.T) {
-	// WHEN
-	jobConfig, err := tester.ReadJobConfig("./../../../../../prow/jobs/incubator/documentation-component/documentation-component-governance.yaml")
-	// THEN
-	require.NoError(t, err)
-
-	periodics := jobConfig.Periodics
-	assert.Len(t, periodics, 1)
-
-	expName := "documentation-component-governance-nightly"
-	actualPeriodic := tester.FindPeriodicJobByName(periodics, expName)
-	require.NotNil(t, actualPeriodic)
-	assert.Equal(t, expName, actualPeriodic.Name)
-	assert.True(t, actualPeriodic.Decorate)
-	assert.Equal(t, "0 1 * * 1-5", actualPeriodic.Cron)
-	tester.AssertThatHasPresets(t, actualPeriodic.JobBase, tester.PresetDindEnabled)
-	tester.AssertThatHasExtraRefs(t, actualPeriodic.JobBase.UtilityConfig, []string{"test-infra", "documentation-component"})
-	assert.Equal(t, tester.ImageBootstrapLatest, actualPeriodic.Spec.Containers[0].Image)
-	assert.Equal(t, []string{tester.GovernanceScriptDir}, actualPeriodic.Spec.Containers[0].Command)
-	repositoryDirArg := fmt.Sprintf("%s/documentation-component", tester.KymaIncubatorDir)
-	assert.Equal(t, []string{"--repository", "documentation-component", "--repository-org", "kyma-incubator", "--repository-dir", repositoryDirArg, "--full-validation", "true"}, actualPeriodic.Spec.Containers[0].Args)
-}
