+++ prow/jobs/incubator/documentation-component/documentation-component.yaml
@@ -11,7 +11,7 @@ job_template: &job_template
       path_alias: github.com/kyma-project/test-infra
   spec:
     containers:
-      - image: eu.gcr.io/kyma-project/prow/test-infra/buildpack-node:v20181130-b28250b
+      - image: eu.gcr.io/kyma-project/test-infra/buildpack-node:v20190724-d41df0f
         securityContext:
           privileged: true
         command:
@@ -28,10 +28,74 @@ job_labels_template: &job_labels_template
 
 presubmits: # runs on PRs
   kyma-incubator/documentation-component:
-    - name: pre-master-documentation-component
-      branches:
-        - ^master$
+    - name: pre-documentation-component
       labels:
         <<: *job_labels_template
         preset-build-pr: "true"
       <<: *job_template
+    - name: pre-documentation-component-governance
+      skip_report: false
+      decorate: true
+      max_concurrency: 10
+      labels:
+        <<: *job_labels_template
+        preset-build-pr: "true"
+      run_if_changed: "milv.config.yaml|.md$"
+      path_alias: github.com/kyma-incubator/documentation-component
+      extra_refs:
+        - org: kyma-project
+          repo: test-infra
+          base_ref: master
+          path_alias: github.com/kyma-project/test-infra
+      spec:
+        containers:
+          - image: eu.gcr.io/kyma-project/prow/test-infra/bootstrap:v20181121-f3ea5ce
+            securityContext:
+              privileged: true
+            command:
+              - "/home/prow/go/src/github.com/kyma-project/test-infra/prow/scripts/governance.sh"
+            args:
+              - "--repository"
+              - "documentation-component"
+              - "--repository-org"
+              - "kyma-incubator"
+            resources:
+              requests:
+                memory: 10Mi
+                cpu: 100m
+
+periodics: # runs periodic on master
+- name: documentation-component-governance-nightly
+  decorate: true
+  cron: "0 1 * * 1-5" # “At 01:00 on every day-of-week from Monday through Friday.”
+  labels:
+    preset-dind-enabled: "true"
+  extra_refs:
+    - org: kyma-project
+      repo: test-infra
+      base_ref: master
+      path_alias: github.com/kyma-project/test-infra
+    - org: kyma-incubator
+      repo: documentation-component
+      base_ref: master
+      path_alias: github.com/kyma-incubator/documentation-component
+  spec:
+    containers:
+      - image: eu.gcr.io/kyma-project/prow/test-infra/bootstrap:v20181121-f3ea5ce
+        securityContext:
+          privileged: true
+        command:
+          - "/home/prow/go/src/github.com/kyma-project/test-infra/prow/scripts/governance.sh"
+        args:
+          - "--repository"
+          - "documentation-component"
+          - "--repository-org"
+          - "kyma-incubator"
+          - "--repository-dir"
+          - "/home/prow/go/src/github.com/kyma-incubator/documentation-component"
+          - "--full-validation"
+          - "true"
+        resources:
+          requests:
+            memory: 10Mi
+            cpu: 100m
